Repackage AppTheme
