+++ app/src/main/java/com/OxGames/Pluvia/utils/application/AppTheme.kt
@@ -1,0 +1,51 @@
+package com.OxGames.Pluvia.utils.application
+
+import com.OxGames.Pluvia.PrefManager
+import com.OxGames.Pluvia.enums.AppTheme
+import com.materialkolor.PaletteStyle
+import kotlin.properties.ReadWriteProperty
+import kotlin.reflect.KProperty
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+
+/**
+ * Referenced from https://github.com/fvilarino/App-Theme-Compose-Sample
+ */
+
+interface IAppTheme {
+    val themeFlow: StateFlow<AppTheme>
+    var currentTheme: AppTheme
+    val paletteFlow: StateFlow<PaletteStyle>
+    var currentPalette: PaletteStyle
+}
+
+class AppThemeImpl : IAppTheme {
+
+    override val themeFlow: MutableStateFlow<AppTheme> = MutableStateFlow(PrefManager.appTheme)
+
+    override var currentTheme: AppTheme by AppThemeDelegate()
+
+    override val paletteFlow: MutableStateFlow<PaletteStyle> = MutableStateFlow(PrefManager.appThemePalette)
+
+    override var currentPalette: PaletteStyle by AppPaletteDelegate()
+
+    inner class AppThemeDelegate : ReadWriteProperty<Any, AppTheme> {
+
+        override fun getValue(thisRef: Any, property: KProperty<*>): AppTheme = PrefManager.appTheme
+
+        override fun setValue(thisRef: Any, property: KProperty<*>, value: AppTheme) {
+            themeFlow.value = value
+            PrefManager.appTheme = value
+        }
+    }
+
+    inner class AppPaletteDelegate : ReadWriteProperty<Any, PaletteStyle> {
+
+        override fun getValue(thisRef: Any, property: KProperty<*>): PaletteStyle = PrefManager.appThemePalette
+
+        override fun setValue(thisRef: Any, property: KProperty<*>, value: PaletteStyle) {
+            paletteFlow.value = value
+            PrefManager.appThemePalette = value
+        }
+    }
+}
